updated excel sample
