+++ AddingDataSources/ExcelFile/server/aspnet/RevealSdk.Server/Program.cs
@@ -1,0 +1,46 @@
+using Reveal.Sdk;
+
+var builder = WebApplication.CreateBuilder(args);
+
+// Add services to the container.
+
+builder.Services.AddControllers().AddReveal( builder =>
+{
+    builder.AddSettings(builder =>
+   {
+       builder.LocalFileStoragePath = "Data";
+   });
+});
+
+// Learn more about configuring Swagger/OpenAPI at https://aka.ms/aspnetcore/swashbuckle
+builder.Services.AddEndpointsApiExplorer();
+builder.Services.AddSwaggerGen();
+
+builder.Services.AddCors(options =>
+{
+  options.AddPolicy("AllowAll",
+    builder => builder.AllowAnyOrigin().AllowAnyHeader().AllowAnyMethod()
+  );
+});
+
+var app = builder.Build();
+
+// Configure the HTTP request pipeline.
+if (app.Environment.IsDevelopment())
+{
+    app.UseSwagger();
+    app.UseSwaggerUI();
+}
+
+app.UseHttpsRedirection();
+
+if (app.Environment.IsDevelopment())
+{
+  app.UseCors("AllowAll");
+}
+
+app.UseAuthorization();
+
+app.MapControllers();
+
+app.Run();
